Move simulate_worker utility out of conftest.py into helpers.py
conftest.py is for pytest fixtures and hooks, not general utilities.
Move simulate_worker to tests/stress/helpers.py and update the import
in test_concurrent_workers.py accordingly. Closes #41.
